Update pm-File / Bugfixing
- Angepasstes pm-File für "neue gevoo HM485-lib"
- Bugfix: Kanal 2-4 wurden fehlerhaft berechnet
- Bugfix: Antwort auf Set_Level mit Rückgabe des Zielwerts.

+++ HBW-LC-Bl4/source/HMWModule.h
@@ -21,7 +21,7 @@
 class HMWDeviceBase {
   public:
 	virtual void setLevel(byte,unsigned int) = 0;  // channel, level
-	virtual unsigned int getLevel(byte) = 0;       // channel, returns level
+	virtual byte getLevel(byte) = 0;       // channel, returns level
 	virtual void processKey(unsigned char, byte) = 0;
 	virtual void readConfig() = 0;         // read config from EEPROM
 };
